+++ RELEASE_NOTES.md
@@ -1,11 +1,29 @@
 Release Notes
 =============
+## 1.6.9
+* WebApp: Slots now inherit user assigned identities from their owning webApp
+
+## 1.6.8
+* SQL Azure and Postgres: `add_firewall_rules` to take list of rules
+
+## 1.6.7
+* Container Groups: Reference Azure container registry credentials.
+* DNS Zone: Support for adding records to existing zones.
+* DNS Zone: zone and record 'depends_on' support.
+* DNS Zone: DNS record 'target_resource' fix to emit correct resource Id.
+* Web App, Functions: Refactored Web App and Functions builders to simplify adding new common properties
 
 ## 1.6.6
+* Azure Firewall: Support for 'link_to_firewall_policy' to link to a builder as well as a resource.
 * Container Groups: Support for 'depends_on' to add dependencies.
 * Functions: Added support for deployment slots
 * KeyVault: Enable VaultUri configuration member for use as output parameter.
 * KeyVault: Fix emitted `enablePurgeProtection`.
+* Storage Account: Add support for data protection policies,
+* Storage Account: Add support for versioning.
+* Virtual Network: Specify the network security group for a subnet.
+* Virtual Network: Subnet support for enabling or disabling Private Link Service Network Policies to allow assigning the IP for a private endpoint connection.
+* Virtual Machine: Added support for Private IP on NIC
 * WebApp: Added support for deployment slots
 
 ## 1.6.5